pyproject.toml
[tool.poetry]
name = "mandi-monitor-bot"
version = "0.1.0"
description = "Telegram bot that alerts Indian shoppers on deep discounts."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.0"
SQLModel = "^0.0.14"
APScheduler = "^3.10.4"
python-amazon-paapi = "==5.*"
playwright = "==1.*"
Flask = "^3.0"
sentry-sdk = "^2.0.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.0.0"
regex = "^2024.7.24"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-mock = "^3.0.0"
ruff = "^0.4.0"
black = "^24.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 